Auto remove  setting
