+++ lib/composer.js
@@ -1,10 +1,9 @@
 /** @babel */
 
 import _ from 'lodash'
-import { shell } from 'electron'
 import fs from 'fs-plus'
 import path from 'path'
-import { getEditorDetails, heredoc, replacePropertiesInString } from './werkzeug'
+import { getEditorDetails, replacePropertiesInString } from './werkzeug'
 import minimatch from 'minimatch'
 import glob from 'glob'
 import { CompositeDisposable, Disposable } from 'atom'
@@ -29,31 +28,6 @@ export default class Composer extends Disposable {
     if (!state.builder) {
       latex.log.warning(`No registered LaTeX builder can process ${filePath}.`)
       return state
-    }
-
-    if (Object.getPrototypeOf(state.builder).constructor.name === 'TexifyBuilder') {
-      // -------------------------------------------------------------
-      // TODO: Remove this whole block when texify support is removed.
-      // -------------------------------------------------------------
-      const message = `LaTeX: The texify builder has been deprecated`
-      const description = heredoc(`
-        Support for the \`texify\` builder has been deprecated in favor of \`latexmk\`, and will be
-        removed soon.`)
-
-      const title = 'How to use latexmk with MiKTeX'
-      const url = 'https://github.com/thomasjo/atom-latex/wiki/Using-latexmk-with-MiKTeX'
-      const openUrl = (event) => {
-        // NOTE: Horrible hack due to a bug in atom/notifications module...
-        const element = event.target.parentElement.parentElement.parentElement.parentElement
-        const notification = element.getModel()
-        notification.dismiss()
-
-        shell.openExternal(url)
-      }
-
-      atom.notifications.addWarning(message, {
-        dismissable: true, description, buttons: [{ text: title, onDidClick: openUrl }]
-      })
     }
 
     state.jobnames = state.builder.getJobNamesFromMagic(state.rootFilePath)
