+++ lib/config-migrator.js
@@ -1,5 +1,6 @@
 /** @babel */
 
+import { shell } from 'electron'
 import _ from 'lodash'
 import { heredoc } from './werkzeug'
 
@@ -8,6 +9,38 @@ export default function checkConfigAndMigrate () {
   checkCleanExtensions()
   checkOpenerSetting()
   checkMasterFileSearchSetting()
+  checkBuilder()
+}
+
+function checkBuilder () {
+  const builder = atom.config.get('latex.builder')
+  if (!builder) return
+
+  atom.config.unset('latex.builder')
+  if (builder !== 'texify') return
+
+  // --------------------------------------------------
+  // TODO: Remove this whole block after a grace period
+  // --------------------------------------------------
+  const message = `LaTeX: The texify builder has been deprecated`
+  const description = heredoc(`
+    Support for the \`texify\` builder has been deprecated in favor of \`latexmk\`,
+    and has been removed.`)
+
+  const title = 'How to use latexmk with MiKTeX'
+  const url = 'https://github.com/thomasjo/atom-latex/wiki/Using-latexmk-with-MiKTeX'
+  const openUrl = (event) => {
+    // NOTE: Horrible hack due to a bug in atom/notifications module...
+    const element = event.target.parentElement.parentElement.parentElement.parentElement
+    const notification = element.getModel()
+    notification.dismiss()
+
+    shell.openExternal(url)
+  }
+
+  atom.notifications.addWarning(message, {
+    dismissable: true, description, buttons: [{ text: title, onDidClick: openUrl }]
+  })
 }
 
 function checkMasterFileSearchSetting () {
